try compiler wrapper for emcc and emc++
